make the vault kv backup type more specific
this is because, we clearly know the type at these
levels, though we cannot know the type beyond this
level since the value of the secret could be
simple flat key-value pairs, or a complex nested
JSON. but we surely know that it will be a map /
object with the first level being a string and
we already know that the secret path is going to
be a string. so, with all this in mind, we can
make it more specific

I realized this when working on the
vault-kv-restore tool where we need to parse the
vault kv backup JSON file and I realized that I
don't have to type assert map[string]interface{}
to map[string]map[string]interface{} and do that
conversion. I would rather use
map[string]map[string]interface{} directly while
reading and parsing the kv backup JSON file. hence
this change on the JSON file production / creation
also

Signed-off-by: Karuppiah Natarajan <karuppiah7890@gmail.com>